В имя тестового пакета добавлено слово check
