+++ tests/com.e1c.v8codestyle.md.itests/src/com/e1c/v8codestyle/internal/md/itests/CheckMd.java
@@ -78,4 +78,5 @@ public abstract class CheckMd
         });
         waitForDD(dtProject);
     }
+
 }
