+++ tests/com.e1c.v8codestyle.md.itests/src/com/e1c/v8codestyle/md/check/itests/CommonModuleNameClientServerTest.java
@@ -11,7 +11,7 @@
  *     1C-Soft LLC - initial API and implementation
  *     Aleksandr Kapralov - issue #14
  *******************************************************************************/
-package com.e1c.v8codestyle.md.itests;
+package com.e1c.v8codestyle.md.check.itests;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
@@ -22,33 +22,33 @@ import com._1c.g5.v8.dt.core.platform.IDtProject;
 import com._1c.g5.v8.dt.metadata.mdclass.ReturnValuesReuse;
 import com._1c.g5.v8.dt.validation.marker.Marker;
 import com.e1c.v8codestyle.internal.md.itests.CheckMd;
-import com.e1c.v8codestyle.md.check.CommonModuleNameClient;
+import com.e1c.v8codestyle.md.check.CommonModuleNameClientServer;
 import com.e1c.v8codestyle.md.check.CommonModuleType;
 
 /**
- * Tests for {@link CommonModuleNameClient} check.
+ * Tests for {@link CommonModuleNameClientServer} check.
  *
  * @author Dmitriy Marmyshev
  *
  */
-public class CommonModuleNameClientTest
+public class CommonModuleNameClientServerTest
     extends CheckMd
 {
 
-    private static final String CHECK_ID = "common-module-name-client";
+    private static final String CHECK_ID = "common-module-name-client-server";
 
     private static final String PROJECT_NAME = "CommonModuleName";
 
     private static final String MODULE_DEFAULT_FQN = "CommonModule.CommonModuleName";
 
     @Test
-    public void testCommonModuleNameClient() throws Exception
+    public void testCommonModuleNameClientServer() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT, ReturnValuesReuse.DONT_USE,
-            null);
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
+            ReturnValuesReuse.DONT_USE, null);
 
         long id = getTopObjectIdByFqn(MODULE_DEFAULT_FQN, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
@@ -56,15 +56,15 @@ public class CommonModuleNameClientTest
     }
 
     @Test
-    public void testCommonModuleNameClientCorrect() throws Exception
+    public void testCommonModuleNameClientServerCorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        String fqn = "CommonModule.CommonModuleClient";
+        String fqn = "CommonModule.CommonModuleClientServer";
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT, ReturnValuesReuse.DONT_USE,
-            fqn);
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
+            ReturnValuesReuse.DONT_USE, fqn);
 
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
@@ -72,15 +72,15 @@ public class CommonModuleNameClientTest
     }
 
     @Test
-    public void testCommonModuleNameClientWithPostfixCorrect() throws Exception
+    public void testCommonModuleNameClientServerWithPostfixCorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        String fqn = "CommonModule.CommonModuleClientPredefined";
+        String fqn = "CommonModule.CommonModuleClientServerPredefined";
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT, ReturnValuesReuse.DONT_USE,
-            fqn);
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
+            ReturnValuesReuse.DONT_USE, fqn);
 
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
@@ -88,15 +88,15 @@ public class CommonModuleNameClientTest
     }
 
     @Test
-    public void testCommonModuleNameClientWithPrefixIncorrect() throws Exception
+    public void testCommonModuleNameClientServerWithPrefixIncorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        String fqn = "CommonModule.ClientCommonModule";
+        String fqn = "CommonModule.ClientServerCommonModule";
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT, ReturnValuesReuse.DONT_USE,
-            fqn);
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
+            ReturnValuesReuse.DONT_USE, fqn);
 
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
@@ -104,12 +104,12 @@ public class CommonModuleNameClientTest
     }
 
     @Test
-    public void testCommonModuleNameClientReturnValueReuseCorrect() throws Exception
+    public void testCommonModuleNameClientServerReturnValueReuseCorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT,
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
             ReturnValuesReuse.DURING_SESSION, null);
 
         long id = getTopObjectIdByFqn(MODULE_DEFAULT_FQN, dtProject);
