+++ tests/com.e1c.v8codestyle.md.itests/src/com/e1c/v8codestyle/md/check/itests/CommonModuleNameClientTest.java
@@ -11,7 +11,7 @@
  *     1C-Soft LLC - initial API and implementation
  *     Aleksandr Kapralov - issue #14
  *******************************************************************************/
-package com.e1c.v8codestyle.md.itests;
+package com.e1c.v8codestyle.md.check.itests;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
@@ -22,33 +22,33 @@ import com._1c.g5.v8.dt.core.platform.IDtProject;
 import com._1c.g5.v8.dt.metadata.mdclass.ReturnValuesReuse;
 import com._1c.g5.v8.dt.validation.marker.Marker;
 import com.e1c.v8codestyle.internal.md.itests.CheckMd;
-import com.e1c.v8codestyle.md.check.CommonModuleNameClientServer;
+import com.e1c.v8codestyle.md.check.CommonModuleNameClient;
 import com.e1c.v8codestyle.md.check.CommonModuleType;
 
 /**
- * Tests for {@link CommonModuleNameClientServer} check.
+ * Tests for {@link CommonModuleNameClient} check.
  *
  * @author Dmitriy Marmyshev
  *
  */
-public class CommonModuleNameClientServerTest
+public class CommonModuleNameClientTest
     extends CheckMd
 {
 
-    private static final String CHECK_ID = "common-module-name-client-server";
+    private static final String CHECK_ID = "common-module-name-client";
 
     private static final String PROJECT_NAME = "CommonModuleName";
 
     private static final String MODULE_DEFAULT_FQN = "CommonModule.CommonModuleName";
 
     @Test
-    public void testCommonModuleNameClientServer() throws Exception
+    public void testCommonModuleNameClient() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
-            ReturnValuesReuse.DONT_USE, null);
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT, ReturnValuesReuse.DONT_USE,
+            null);
 
         long id = getTopObjectIdByFqn(MODULE_DEFAULT_FQN, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
@@ -56,15 +56,15 @@ public class CommonModuleNameClientServerTest
     }
 
     @Test
-    public void testCommonModuleNameClientServerCorrect() throws Exception
+    public void testCommonModuleNameClientCorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        String fqn = "CommonModule.CommonModuleClientServer";
+        String fqn = "CommonModule.CommonModuleClient";
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
-            ReturnValuesReuse.DONT_USE, fqn);
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT, ReturnValuesReuse.DONT_USE,
+            fqn);
 
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
@@ -72,15 +72,15 @@ public class CommonModuleNameClientServerTest
     }
 
     @Test
-    public void testCommonModuleNameClientServerWithPostfixCorrect() throws Exception
+    public void testCommonModuleNameClientWithPostfixCorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        String fqn = "CommonModule.CommonModuleClientServerPredefined";
+        String fqn = "CommonModule.CommonModuleClientPredefined";
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
-            ReturnValuesReuse.DONT_USE, fqn);
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT, ReturnValuesReuse.DONT_USE,
+            fqn);
 
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
@@ -88,15 +88,15 @@ public class CommonModuleNameClientServerTest
     }
 
     @Test
-    public void testCommonModuleNameClientServerWithPrefixIncorrect() throws Exception
+    public void testCommonModuleNameClientWithPrefixIncorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        String fqn = "CommonModule.ClientServerCommonModule";
+        String fqn = "CommonModule.ClientCommonModule";
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
-            ReturnValuesReuse.DONT_USE, fqn);
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT, ReturnValuesReuse.DONT_USE,
+            fqn);
 
         long id = getTopObjectIdByFqn(fqn, dtProject);
         Marker marker = getFirstMarker(CHECK_ID, id, dtProject);
@@ -104,12 +104,12 @@ public class CommonModuleNameClientServerTest
     }
 
     @Test
-    public void testCommonModuleNameClientServerReturnValueReuseCorrect() throws Exception
+    public void testCommonModuleNameClientReturnValueReuseCorrect() throws Exception
     {
         IDtProject dtProject = openProjectAndWaitForValidationFinish(PROJECT_NAME);
         assertNotNull(dtProject);
 
-        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT_SERVER,
+        updateCommonModule(dtProject, MODULE_DEFAULT_FQN, CommonModuleType.TYPE_CLIENT,
             ReturnValuesReuse.DURING_SESSION, null);
 
         long id = getTopObjectIdByFqn(MODULE_DEFAULT_FQN, dtProject);
