+++ tests/com.e1c.v8codestyle.md.itests/src/com/e1c/v8codestyle/md/check/itests/ConfigurationDataLockTest.java
@@ -1,7 +1,7 @@
 /**
  *
  */
-package com.e1c.v8codestyle.md.itests;
+package com.e1c.v8codestyle.md.check.itests;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
